requirements.txt
Flask~=2.2.2
Werkzeug~=2.2.2
SQLAlchemy~=1.4.44
psycopg2-binary~=2.9.5
#requests~=2.28.1
SQLAlchemy-Utils~=0.38.3
flask_restful~=0.3.9
#flask_login~=0.6.2
#flask_sqlalchemy~=3.0.2
#flask_jwt_extended~=4.4.4
#pytest~=7.2.0
#psycopg2~=2.9.5
paho-mqtt~=1.6.1
